long way of disabling collision in main()
This is the first version of the temporary disabling the collision check by detecting the all Zero State of the robot in the SRS emulator. An if-else statement is used with scopes, in which the related settings were done explicitly but the overall code is long. Next upload will use the helper methods to reduce the within-main code