collision disable start algorithm with helper
Temporary collision disabling/enabling is done with a helper method out of main() to reduce the length of the code for detecting the start state of the robot in the emulator (whether all Zero State -only in Emulator start of ros_server application- OR non-Zero State -which is the real case-). After this development, the helper method is going to be integrated into ecoswash_moveit_helper.cpp source file. The execution hanging problem still occurs sometimes and randomly.